Refactor position are always integer value

+++ Outdated-tank/src/main/scala/models/Position.scala
@@ -1,3 +1,3 @@
 package isc.game.outdatedtank.models
 
-case class Position(var x: Double, var y: Double)
+case class Position(var x: Int, var y: Int)
